requirements.txt
Pillow==2.8.1
PyYAML==3.11
scruffy==0.3
